RSC: Move vite plugins into their own files
