+++ packages/vite/src/rsc/rscBuildAnalyze.ts
@@ -4,7 +4,7 @@ import { getPaths } from '@redwoodjs/project-config'
 
 import { onWarn } from '../lib/onWarn.js'
 
-import { rscAnalyzePlugin } from './rscVitePlugins.js'
+import { rscAnalyzePlugin } from '../plugins/vite-plugin-rsc-analyze.js'
 
 /**
  * RSC build. Step 1.
